refactor extensions
